robot-dashboard/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5
websocket-client==1.7.0
watchdog==3.0.0
numpy==1.26.2
python-socketio==5.10.0